refactor rename blueprintsSlice -> modsSlice
