+++ src/extensionConsole/pages/mods/GetStartedView.stories.tsx
@@ -19,9 +19,9 @@ import React from "react";
 import { type ComponentStory, type ComponentMeta } from "@storybook/react";
 import { configureStore } from "@reduxjs/toolkit";
 import { persistReducer } from "redux-persist";
-import blueprintsSlice, {
+import modsSlice, {
   persistBlueprintsConfig,
-} from "@/extensionConsole/pages/mods/blueprintsSlice";
+} from "@/extensionConsole/pages/mods/modsSlice";
 import { appApi } from "@/services/api";
 import GetStartedView from "@/extensionConsole/pages/mods/GetStartedView";
 import { Provider } from "react-redux";
@@ -39,10 +39,7 @@ export default {
 function optionsStore(initialState?: any) {
   return configureStore({
     reducer: {
-      blueprints: persistReducer(
-        persistBlueprintsConfig,
-        blueprintsSlice.reducer
-      ),
+      blueprints: persistReducer(persistBlueprintsConfig, modsSlice.reducer),
       recipes: recipesSlice.reducer,
       auth: persistReducer(persistAuthConfig, authSlice.reducer),
       [appApi.reducerPath]: appApi.reducer,
