+++ src/extensionConsole/pages/mods/ModsPageLayout.test.tsx
@@ -21,7 +21,7 @@ import ModsPageLayout from "@/extensionConsole/pages/mods/ModsPageLayout";
 import { type Mod } from "@/types/modTypes";
 import { waitForEffect } from "@/testUtils/testHelpers";
 import { act, screen } from "@testing-library/react";
-import blueprintsSlice from "@/extensionConsole/pages/mods/blueprintsSlice";
+import modsSlice from "@/extensionConsole/pages/mods/modsSlice";
 import userEvent from "@testing-library/user-event";
 import { authSlice } from "@/auth/authSlice";
 import { mockCachedUser, mockLoadingUser } from "@/testUtils/userMock";
@@ -206,9 +206,7 @@ describe("Serializable Data Test", () => {
     render(<ModsPageLayout mods={mods} />, {
       setupRedux(dispatch) {
         dispatch(
-          blueprintsSlice.actions.setSearchQuery(
-            (() => {}) as unknown as string
-          )
+          modsSlice.actions.setSearchQuery((() => {}) as unknown as string)
         );
       },
     });
